move manager.run tests -> backend abstract
